+++ src/drivers/cmakeLegacyDriver.ts
@@ -19,6 +19,7 @@ import * as nls from 'vscode-nls';
 import { BuildPreset, ConfigurePreset, getValue, TestPreset } from '@cmt/preset';
 import { CodeModelContent } from './codeModel';
 import { ConfigureTrigger } from '@cmt/cmakeProject';
+import { onConfigureSettingsChange } from '@cmt/ui/util';
 
 nls.config({ messageFormat: nls.MessageFormat.bundle, bundleFormat: nls.BundleFormat.standalone })();
 const localize: nls.LocalizeFunc = nls.loadMessageBundle();
@@ -45,7 +46,7 @@ export class CMakeLegacyDriver extends CMakeDriver {
     private _needsReconfigure = true;
     async doConfigureSettingsChange(): Promise<void> {
         this._needsReconfigure = true;
-        await util.onConfigureSettingsChange();
+        await onConfigureSettingsChange();
     }
     async checkNeedsReconfigure(): Promise<boolean> {
         return this._needsReconfigure;
